requirements.txt
aiofiles==23.1.0
starlette-admin[i18n]==0.8.0.dev0
itsdangerous==2.1.2
uvicorn==0.21.1
sqlalchemy-file==0.4.0
fasteners==0.18
mongoengine==0.27.0
Pillow==9.5.0
Markdown==3.4.3
sqlmodel==0.0.8
pydantic[email]==1.10.7
PyMySQL==1.0.3
odmantic==0.9.2
